Organized PyTest fixtures into submodules to separate between ssh testing and (future) base testing
