+++ tests/switcheroo/ssh/end_to_end/test_publish_and_read.py
@@ -3,7 +3,7 @@ from getpass import getuser
 from pathlib import Path
 from paramiko import SSHClient, AutoAddPolicy, RSAKey
 import pytest
-from tests.switcheroo_ssh.server import Server
+from tests.switcheroo.ssh.server import Server
 from switcheroo.ssh.data_org.publisher.s3 import S3KeyPublisher
 from switcheroo.ssh.data_org.retriever.s3 import S3KeyRetriever
 from switcheroo import paths
