+++ tests/switcheroo/ssh/unit/publisher/test_local_publisher.py
