+++ tests/switcheroo/ssh/unit/publisher/test_s3_publisher.py
